LayoutTests/fast/js/sputnik/Conformance/11_Expressions/11.4_Unary_Operators/11.4.4_Prefix_Increment_Operator/S11.4.4_A2.1_T3-expected.txt

S11.4.4_A2.1_T3

PASS ReferenceError: Prefix ++ operator applied to value that is not a reference.

TEST COMPLETE
